Debugging tree matching

+++ GasFlowTools.py
@@ -282,7 +282,7 @@ def match_tree(mcut,snapidxmin=0):
                 print(f'getting tree data = {time.time()-t0halo:.2f}')
                 isub_treedata=snap_tree_catalogue.loc[isub_match,fields_tree]
                 print(f'putting tree data in array = {time.time()-t0halo:.2f}')
-                catalogue_subhalo.iloc[isub,fields_tree_index]=(isub_treedata.values).astype(int)
+                catalogue_subhalo.iloc[isub,fields_tree_index]=isub_treedata.values
             else:
                 logging.info(f'Warning: could not match subhalo {iisub} at ({isub_com[0]:.2f},{isub_com[1]:.2f},{isub_com[2]:.2f}) cMpc')
                 pass
